+++ ci/build-1-autotools.sh
@@ -10,8 +10,8 @@ fi
 # We keep our own list of mirrors because https://ftpmirror.gnu.org is
 # unreliable (frequent errors from selected mirror).
 MIRRORS=(
-    https://mirror.cs.odu.edu/gnu
     https://mirrors.ocf.berkeley.edu/gnu
+    https://mirror.cs.odu.edu/gnu
     https://ftp.gnu.org/gnu
 )
 
@@ -21,6 +21,7 @@ LIBTOOL_REL=libtool/libtool-2.5.4.tar.gz
 
 PREFIX=$(pwd)/ci/build
 PATH=$(pwd)/ci/build/bin:$PATH
+KEYRING=$(pwd)/ci/fping-deps.gpg
 
 if [ ! -d ci ]; then
     echo "you must run this in the root fping directory" >&2
@@ -30,28 +31,43 @@ fi
 # remove standard versions
 sudo apt-get remove -qq autoconf automake autotools-dev libtool
 
+# install dependencies
+sudo apt-get install -y gpgv
+
 # prepare build environment
 cd ci
 rm -rf build
 mkdir -p build/src
 cd build/src
 
+mirror_fetch() {
+    local relpath="$1"
+    for mirror in "${MIRRORS[@]}"; do
+        local url="$mirror/$relpath"
+        if wget -t 3 "$url"; then
+            return 0
+        fi
+    done
+    return 1
+}
+
 install_release() {
-    local relpath=$1
+    local relpath="$1"
     local file=$(basename "$relpath")
     local dir="${file%%.tar.*}"
 
-    local success=0
-    for mirror in "${MIRRORS[@]}"; do
-        local url="$mirror/$relpath"
-        if wget -t 3 -O "$file" "$url"; then
-            success=1
-            break
-        fi
-    done
-
-    if [ $success -eq 0 ]; then
+    if ! mirror_fetch "$relpath"; then
         echo "Failed to download $relpath from any mirror" >&2
+        exit 1
+    fi
+
+    if ! mirror_fetch "$relpath.sig"; then
+        echo "Failed to download $relpath.sig from any mirror" >&2
+        exit 1
+    fi
+
+    if ! gpgv --keyring "$KEYRING" "$file.sig" "$file"; then
+        echo "GPG verification failed for $file"
         exit 1
     fi
 
@@ -61,7 +77,7 @@ install_release() {
         ./configure --prefix=$PREFIX
         make install
     )
-    rm "$file"
+    rm "$file" "$file.sig"
 }
 
 # autoconf