+++ CHANGELOG.md
@@ -43,6 +43,7 @@ fping 5.5-rc1 (2025-12-21)
 - Updated autoconf from 2.71 to 2.72
 - Updated automake from 1.16.5 to 1.18.1
 - Updated libtool from 2.4.6 to 2.5.4
+- Implemented verification of autotools tarballs in Github actions.
 
 fping 5.4 (2025-04-19)
 ======================